# misc/api-doc.yaml
...
openapi: 3.0.0
info:
  title: 專題 api
  description: |
    This is for project app api
  version: 1.0.0
servers:
- url: http://localhost:3000
...
paths:
...
  /posts:
    get:
      tags:
      - post
      summary: 取得 文章
      description: 取得文章、發文者資訊
      operationId: getPosts
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
              - memNum: 1
                memName: KimoJi
                memPhoto: null
                memIdentify: 0
                postNum: 2
                postCreateAt: 2021-09-02T02:28:46.000Z
                postUpdateAt: 2021-11-29T06:00:25.000Z
                postContent: 哈哈哈哈哈哈哈哈哈哈!!!
                commentNumber: 2
                likeNumber: 1
              - memNum: 3
                memName: 野原新之助
                memPhoto: null
                memIdentify: 1
                postNum: 1
                postCreateAt: 2021-09-02T02:27:17.000Z
                postUpdateAt: 2021-11-25T07:06:30.000Z
                postContent: 我喜歡黃色！！！
                commentNumber: 1
                likeNumber: 2
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /posts/create:
    post:
      tags:
      - post
      summary: 新增 文章
      description: RequestBody各項參數說明 <ul> <li>memNum：用戶序號</li> <li>postContent：文章內容</li> </ul>
      operationId: createPost
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              postData:
                memNum: 1
                postContent: 這麼好用的網站，還不用爆！
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 文章新增成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /posts/update/{postNum}:
    put:
      tags:
      - post
      summary: 編輯 文章
      description: RequestBody各項參數說明 <ul> <li>postContent：文章內容</li> </ul>
      operationId: updatePost
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              postContent: 寫API文件寫到快瘋掉
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 文章編輯成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /posts/delete/{postNum}:
    delete:
      tags:
      - post
      summary: 刪除 文章
      operationId: deletePost
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 文章刪除成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...
  /posts/liked/{postNum}:
    post:
      tags:
      - post
      summary: 文章按讚
      description: RequestBody各項參數說明 <ul> <li>memNum：按讚者序號</li> </ul>
      operationId: postLike
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 1
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功為貼文按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /posts/liked/{postNum}/{memNum}:
    delete:
      tags:
      - post
      summary: 取消 文章按讚
      operationId: removePostLike
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功取消貼文按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
    get:
      tags:
      - post
      summary: 檢查用戶是否按讚
      operationId: postLikeVerify
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 已按讚
        "400":
          description: 找不到資訊
          content:
            application/json:
              example:
                message: 未按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /posts/report/{postNum}:
    post:
      tags:
      - post
      summary: 檢舉文章
      description: RequestBody各項參數說明 <ul> <li>memNum：檢舉者序號</li> </ul>
      operationId: postReport
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 1
              reportReason: 檢舉測試
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功檢舉貼文
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...
  /products:
    get:
      tags:
      - product
      summary: 取得 商品
      description: 取得全部商品資訊 或 取得篩選過後的商品資訊(模糊比對)
      operationId: getProducts
      parameters:
      - name: proCompany
        in: query
        description: 商品發行公司
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proBigItem
        in: query
        description: 商品分類大項
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proSmallItem
        in: query
        description: 商品分類小項
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proKind
        in: query
        description: 主約or附約
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proPeriod
        in: query
        description: 終身or定期
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proName
        in: query
        description: 商品名稱(中英文皆可)
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: NTDD
      - name: proRemark
        in: query
        description: 商品備註
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proFeature
        in: query
        description: 商品特色
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proContent
        in: query
        description: 商品保障項目
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: proStatus
        in: query
        description: 商品銷售狀態
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
              - proNum: 1
                proCompany: 南山人壽
                proBigItem: '[''壽險保障商品'', ''醫療保障商品'']'
                proSmallItem: '[''定期壽險-台幣'', ''重大疾病與重大傷病'', ''癌症醫療'']'
                proKind: 主約
                proPeriod: 定期
                proName: 南山人壽新活力康祥定期健康保險(NTDD)
                proRemark: ""
                proFeature: '[''投保年齡：15足歲~40歲'', ''7項重大疾病，輕/重度均有保障'', ''加強人生黃金期的階段性保障'']'
                proContent: '[''身故給付'', ''完全失能給付'', ''重大疾病'', ''豁免保費'']'
                proDM: https://www.nanshanlife.com.tw/NanshanWeb/file/DOCUMENT/7341
                proTerms: https://www.nanshanlife.com.tw/NanshanWeb/file/DOCUMENT/7361
                proStatus: 現售
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /products/liked/{proNum}:
    post:
      tags:
      - product
      summary: 商品按讚
      description: RequestBody各項參數說明 <ul> <li>memNum：按讚者序號</li> </ul>
      operationId: productLike
      parameters:
      - name: proNum
        in: path
        description: 商品序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 30
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 1
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功為商品按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /products/liked/{proNum}/{memNum}:
    delete:
      tags:
      - product
      summary: 取消 商品按讚
      operationId: removeProductLike
      parameters:
      - name: proNum
        in: path
        description: 商品序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 43
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功取消商品按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
    get:
      tags:
      - product
      summary: 檢查用戶是否按讚
      operationId: productLikeVerify
      parameters:
      - name: proNum
        in: path
        description: 商品序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 30
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 已按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...
  /collections/posts/{postNum}:
    get:
      tags:
      - collect
      summary: 取得 文章被收藏數
      operationId: getPostCollect
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                Number: 1
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
    post:
      tags:
      - collect
      summary: 收藏文章
      description: RequestBody各項參數說明 <ul> <li>memNum：收藏者序號</li> </ul>
      operationId: postCollect
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 2
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功收藏了貼文
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /collections/posts/{postNum}/{memNum}:
    delete:
      tags:
      - collect
      summary: 取消 收藏文章
      operationId: removePostCollect
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      - name: memNum
        in: path
        description: 收藏者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功取消貼文收藏
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...
  /collections/products/{proNum}:
    get:
      tags:
      - collect
      summary: 取得 商品被收藏數
      operationId: getProductCollect
      parameters:
      - name: proNum
        in: path
        description: 商品序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 16
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                Number: 1
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
    post:
      tags:
      - collect
      summary: 收藏商品
      description: RequestBody各項參數說明 <ul> <li>memNum：收藏者序號</li> </ul>
      operationId: productCollect
      parameters:
      - name: proNum
        in: path
        description: 商品序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 16
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 2
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功收藏了貼文
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /collections/products/{proNum}/{memNum}:
    delete:
      tags:
      - collect
      summary: 取消 收藏商品
      operationId: removeProductCollect
      parameters:
      - name: proNum
        in: path
        description: 商品序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 16
      - name: memNum
        in: path
        description: 收藏者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功取消貼文收藏
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...